Refactoring of Authentication code and DbInitializer
A vulnerability in the JWT security key generation was removed. The JWT security key is now created based on a randomly generated secret.

The JwtSecurityKey class was replaced with a class called AuthenticationHelper. This class contains methods related to authentication - also some methods which was originally in the TokenController and DbInitializer classes.

There are no static classes in the solution any more. You should generally avoid static classes whenever possible. They cannot be unit tested, and they cannot be faked.